requirements.txt
flask==2.0.1
openai==1.0.0
python-dotenv==0.19.0
requests==2.26.0
gunicorn==20.1.0
PyPDF2==3.0.1
Pillow==9.5.0
werkzeug==2.0.3